make tests using softhsm work with pytest
All the tests worked if started individually, or when started using
nosetest, but pytest had some kind of issue where it calls the tearDown
method for some test while running another test, resulting in
semi-intermittent test failures in this module.

The recommended way to do resource initalisation and cleanup with pytest
is using fixtures. This commit changes this particular module to use
pytest specific functionality, which has the unfortunate effect of
making switching test framework in the future harder.